Reorganized and also broke the code
Reorganized code into sections by satellite rather than task, as reading in all the data in the same section was not working.  Also took out the separation by orbit number in the for loops for reading in the data.

In progress: reading Swarm data from ASCII files.  Might be a good idea to go back to the original organization of the routine and try using parallel for loops instead of regular for and see if that speeds it up.
PROBLEM: dlmread() is returning error message "Internal size mismatch".  Reading in the data is not working.