Routine for importing data
Last version!
Opening ASCII files with Matlab and manually importing the data into the workspace was the best way I found to load the Swarm data.  Trying to actually import it in the script wasn't working, so this routine just takes the imported data from the workspace, organizes it into a structure array, and saves it to a .mat file.